Change the logic of the filename argument to show_refs/show_backrefs:
  * if not specified (i.e. None), spawn xdot as before
  * if ends in '.dot', skip running xdot/dot
  * if ends in '.png', skip xdot, run dot
  * otherwise skip xdot/dot, complain to stdout

Also made it not fail with OSError if 'dot' was not installed.

Also made it create temporary files with tempfile.NamedTemporaryFile.

Also fixed setup.py --build-images and rewrote it to use doctest directly,
since it no longer inserts code in between doctest sections.  As a side
effect the --build-images thing now acts as a (rather small) test suite.