Slight refactoring and testing of edge culling
show_backrefs() stops the graph traversal when it reaches a module
(because that makes the graphs prettier and usually it doesn't matter
where the module is backreferenced: you can treat modules as garbage
collector roots when you're hunting down leaks).

Previously this was hardcoded in a rather ugly way (swap_source_target
was used to distinguish whether the caller was show_refs/show_backrefs).
Now it's parametrized via a new `cull_func` predicate argument to
_show_graph().  If cull_func(node) returns True, then node will appear
in the graph but all graph edges originating at node will be ignored.