Fix Appveyor
At the moment graphviz.portable is not installable (missing checksums),
but graphviz is (only it's not on %PATH% and it took me ages to figure
out where it installs all the stuff).

I expect this to break once a newer graphviz version is out, due to the
hardcoded path etc.  Hopefully the portable package will get fixed by
then.