Python 3.7: gc.get_referrers(a_local_var) is []
The important thing for these tests is that the internal data structures
in objgraph's functions do not retain extra references, so we want to
know that the number of referrers does not increase.  It's less
important to make specific assertions about the actual number of
referrers, which apparently differs across Python versions.

I'd like somebody to explain to me why Python 3.7 no longer includes the
currently executing stack frame in the list of referrers for a local
variable.  Some kind of an optimization?